Move Input/* to Ui/* and alias as $ui – ✨
