+++ app/jsconfig.json
@@ -13,6 +13,7 @@
       "$shared/**/*": ["src/shared/**/*"],
       "$editor/**/*": ["src/editor/**/*"],
       "$testUtils/**/*": ["src/test/test-utils/**/*"],
+      "$ui/**/*": ["src/shared/components/Ui/**/*"],
       "$utils/**/*": ["src/utils/**/*"]
     }
   },
