+++ app/src/auth/components/RegisterPage/index.jsx
@@ -19,7 +19,7 @@ import Button from '../Button'
 import Checkbox from '$shared/components/Checkbox'
 import AuthStep from '../AuthStep'
 import AuthLayout from '../AuthLayout'
-import Text from '$shared/components/Input/Text'
+import Text from '$ui/Text'
 import FormControlLabel from '$shared/components/FormControlLabel'
 import FormControlUnderline from '$shared/components/FormControlUnderline'
 import FormControlErrors from '$shared/components/FormControlErrors'
