+++ app/src/auth/components/ResetPasswordPage/index.jsx
@@ -17,7 +17,7 @@ import Actions from '../Actions'
 import Button from '../Button'
 import AuthStep from '../AuthStep'
 import AuthLayout from '../AuthLayout'
-import Text from '$shared/components/Input/Text'
+import Text from '$ui/Text'
 import FormControlLabel from '$shared/components/FormControlLabel'
 import FormControlUnderline from '$shared/components/FormControlUnderline'
 import FormControlErrors from '$shared/components/FormControlErrors'
