+++ app/src/auth/components/SignupPage/index.jsx
@@ -11,7 +11,7 @@ import Actions from '../Actions'
 import Button from '../Button'
 import AuthStep from '../AuthStep'
 import AuthLayout from '../AuthLayout'
-import Text from '$shared/components/Input/Text'
+import Text from '$ui/Text'
 import FormControlLabel from '$shared/components/FormControlLabel'
 import FormControlUnderline from '$shared/components/FormControlUnderline'
 import FormControlErrors from '$shared/components/FormControlErrors'
