+++ app/src/auth/components/UsernamePasswordLogin/index.jsx
@@ -12,7 +12,7 @@ import AuthPanel from '../AuthPanel'
 import Actions from '../Actions'
 import Button from '../Button'
 import AuthStep from '../AuthStep'
-import Text from '$shared/components/Input/Text'
+import Text from '$ui/Text'
 import FormControlLabel from '$shared/components/FormControlLabel'
 import FormControlUnderline from '$shared/components/FormControlUnderline'
 import FormControlErrors from '$shared/components/FormControlErrors'
