+++ app/src/marketplace/components/MarkdownEditor/index.jsx
@@ -3,7 +3,7 @@
 import React, { useState, useCallback, useEffect } from 'react'
 import cx from 'classnames'
 
-import Text from '$shared/components/Input/Text'
+import Text from '$ui/Text'
 import FormControlErrors from '$shared/components/FormControlErrors'
 import { useLastError, type LastErrorProps } from '$shared/hooks/useLastError'
