+++ app/src/shared/components/Ui/Text/SelectAllOnFocusDecorator.jsx
