+++ app/src/userpages/components/ProfilePage/IdentityHandler/IdentityNameDialog/index.jsx
@@ -6,7 +6,7 @@ import { I18n } from 'react-redux-i18n'
 import ModalPortal from '$shared/components/ModalPortal'
 import Dialog from '$shared/components/Dialog'
 import FormControlLabel from '$shared/components/FormControlLabel'
-import Text from '$shared/components/Input/Text'
+import Text from '$ui/Text'
 
 import styles from './identityNameDialog.pcss'
