+++ app/src/userpages/components/StreamPage/ConfirmCsvImportDialog/index.jsx
@@ -7,9 +7,9 @@ import type { StreamId } from '$shared/flowtype/stream-types'
 import type { CsvUploadState } from '$userpages/flowtype/states/stream-state'
 import ModalPortal from '$shared/components/ModalPortal'
 import Dialog from '$shared/components/Dialog'
-import Select from '$shared/components/Input/Select'
+import Select from '$ui/Select'
 import FormControlLabel from '$shared/components/FormControlLabel'
-import Text from '$shared/components/Input/Text'
+import Text from '$ui/Text'
 import FormControlErrors from '$shared/components/FormControlErrors'
 
 import styles from './confirmCsvImportDialog.pcss'
