+++ app/src/userpages/components/StreamPage/Show/ConfigureView/NewFieldEditor/index.jsx
@@ -4,8 +4,8 @@ import React, { Component } from 'react'
 import { I18n, Translate } from 'react-redux-i18n'
 
 import Button from '$shared/components/Button'
-import Select from '$shared/components/Input/Select'
-import Text from '$shared/components/Input/Text'
+import Select from '$ui/Select'
+import Text from '$ui/Text'
 import type { StreamField } from '$shared/flowtype/stream-types'
 import SplitControl from '$userpages/components/SplitControl'
 import FormControlLabel from '$shared/components/FormControlLabel'
