+++ app/src/userpages/components/StreamPage/Show/ConfigureView/index.jsx
@@ -13,10 +13,10 @@ import type { Stream, StreamId } from '$shared/flowtype/stream-types'
 import type { StoreState } from '$shared/flowtype/store-state'
 import FieldList from '$shared/components/FieldList'
 import FieldItem from '$shared/components/FieldList/FieldItem'
-import Select from '$shared/components/Input/Select'
+import Select from '$ui/Select'
 import { updateEditStreamField, updateEditStream, streamFieldsAutodetect } from '$userpages/modules/userPageStreams/actions'
 import { selectEditedStream, selectFieldsAutodetectFetching, fieldTypes } from '$userpages/modules/userPageStreams/selectors'
-import Text from '$shared/components/Input/Text'
+import Text from '$ui/Text'
 import SplitControl from '$userpages/components/SplitControl'
 import DropdownActions from '$shared/components/DropdownActions'
 import ActionsDropdown from '$shared/components/ActionsDropdown' // look up! lol #naming
