+++ app/src/userpages/components/StreamPage/Show/HistoryView/index.jsx
@@ -13,8 +13,8 @@ import type { CsvUploadState } from '$userpages/flowtype/states/stream-state'
 import { getRange, deleteDataUpTo, uploadCsvFile, confirmCsvFileUpload, updateEditStream } from '$userpages/modules/userPageStreams/actions'
 import { selectDeleteDataError, selectUploadCsvState, selectEditedStream } from '$userpages/modules/userPageStreams/selectors'
 import Button from '$shared/components/Button'
-import Text from '$shared/components/Input/Text'
-import Select from '$shared/components/Input/Select'
+import Text from '$ui/Text'
+import Select from '$ui/Select'
 import FileUpload from '$shared/components/FileUpload'
 import DatePicker from '$shared/components/DatePicker'
 import SvgIcon from '$shared/components/SvgIcon'
