+++ app/src/userpages/components/StreamPage/Show/StatusView/index.jsx
@@ -8,8 +8,8 @@ import type { Stream } from '$shared/flowtype/stream-types'
 import type { StoreState } from '$shared/flowtype/store-state'
 import { updateEditStream } from '$userpages/modules/userPageStreams/actions'
 import { selectEditedStream } from '$userpages/modules/userPageStreams/selectors'
-import Text from '$shared/components/Input/Text'
-import Select from '$shared/components/Input/Select'
+import Text from '$ui/Text'
+import Select from '$ui/Select'
 import FormControlLabel from '$shared/components/FormControlLabel'
 
 import styles from './statusView.pcss'
